Use explicit #to_s to render attribute in form
Since Rails 4.2 html_safe is not used when outputting the attribute in
a text_field which in turn not runs #to_s on it anymore. Because of
this, Rails ends up running join(' ') on the attributes, if they are
arrays which is no good. Using an explicit #to_s makes Arrays render as
comma separated values like before.

This also reversed the options and defaults to make them overrideable.